feat: Analyze A/B test is used to compare the performance of two web pages in a given dataset to see whether the new page leads to more conversion rate than the old page.
The p-value is calculated manually assuming that the null hypothesis is the converted success rate of the new page is equal to that of the old page to see whether the performance of new page is significantly better than the old page. The z-test is also used to check the p-value. At last, use the logistic regression model to investigate possible factors that may affect a individual conversion.